initial commit
Added source files. Keep in mind code may fail when opening up certain views. This is because its trying to pull data that hasn't been populated on pythonanywhere. Need to add a try/except on the code to make sure it runs smoothly as well as check if the pythonanywhere is pointing to the right db